add: new interactive plotting function #7
here, subsets of the timeseries will be plotted based on the user's
specifications, and will show the same subset changing in real time as
each additional pipeline step is performed. a user can also take an
`eyeris` object previously run/saved out and pass it to the generic S3
`plot()` method to obtain a separate plot for each preprocessing step
previously performed on those data.